pyproject.toml
[project]
name = "quizly-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "django>=6.0.3",
  "django-cors-headers>=4.9.0",
  "djangorestframework>=3.16.1",
  "djangorestframework-simplejwt>=5.5.1",
  "google-genai>=1.66.0",
  "python-dotenv>=1.2.2",
  "youtube-transcript-api>=1.2.4",
]

[dependency-groups]
dev = ["ruff>=0.15.5"]

[tool.ruff]
line-length = 79
target-version = "py38"
indent-width = 4

exclude = [".git", "__pycache__", "build", "dist", "venv"]

[tool.ruff.lint]
select = ["E", "F"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
